Minor refactorings to the Date API
Should have parseDuration instead of parsePeriod.
Removed several unnecessary integer casts.

Closed #354

+++ src/main/java/org/mafagafogigante/dungeon/date/Date.java
@@ -17,6 +17,7 @@ import java.io.Serializable;
 public class Date implements Comparable<Date>, Serializable {
 
   private static final long serialVersionUID = Version.MAJOR;
+
   /**
    * The time, in milliseconds.
    */
@@ -135,7 +136,7 @@ public class Date implements Comparable<Date>, Serializable {
    * @return a new Date object
    * @throws IllegalArgumentException if amount is not positive or if unit is null
    */
-  public Date plus(int amount, DungeonTimeUnit unit) {
+  public Date plus(long amount, DungeonTimeUnit unit) {
     if (amount <= 0) {
       throw new IllegalArgumentException("amount must be positive.");
     } else if (unit == null) {
@@ -153,7 +154,7 @@ public class Date implements Comparable<Date>, Serializable {
    * @return a new Date object
    * @throws IllegalArgumentException if amount is not positive or if unit is null
    */
-  public Date minus(int amount, DungeonTimeUnit unit) {
+  public Date minus(long amount, DungeonTimeUnit unit) {
     if (amount <= 0) {
       throw new IllegalArgumentException("amount must be positive.");
     } else if (unit == null) {
